+++ mcramp/LEGACY/geom/plane.py
@@ -9,8 +9,8 @@ import os
 class GPlane(GPrim):
     def __init__(self, width=0, height=0, position=[0, 0, 0], idx=0, ctx=None):
         self.position   = position
-        self.width      = np.float64(width)
-        self.height     = np.float64(height)
+        self.width      = np.float32(width)
+        self.height     = np.float32(height)
         self.idx        = np.uint32(idx)
 
         with open(os.path.join(os.path.dirname(os.path.abspath(__file__)), 'plane.cl'), mode='r') as f:
