+++ mcramp/scat/e_mon.py
@@ -61,4 +61,4 @@ class EMon(SPrim):
     @binning.setter
     def binning(self, val):
         self._binning = np.array((val[0], val[1], val[2], 0.),
-                                 dtype=clarr.vec.double3)
+                                 dtype=clarr.vec.float3)
